requirements.txt
aiohttp==1.0.5
async-timeout==2.0.0
certifi==2018.1.18
chardet==3.0.4
discord.py==0.16.12
fuzzywuzzy==0.16.0
idna==2.6
multidict==4.1.0
pkg-resources==0.0.0
python-Levenshtein==0.12.0
requests==2.18.4
urllib3==1.22
websockets==3.4
